WIP: major-mode based contact list UI to replace old widget based hacktoberfest
While the widget system is powerful, it is arcane and actively hostile
to efficiently locating and updating widgets other than through actual
UI interaction. Rather than fighting the widget API, use major-mode +
font-lock magic to creat pseudo-widgets on the fly. This is both faster
and easier to handle from an internal API point of view.